requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
appdirs==1.4.0            # via setuptools
dnspython==1.15.0
packaging==16.8           # via setuptools
py==1.4.32                # via pytest
pyparsing==2.1.10         # via packaging
pytest==3.0.6
requests==2.13.0
six==1.10.0               # via packaging, setuptools

# The following packages are considered to be unsafe in a requirements file:
# setuptools                # via pytest
